+++ ietfparse/__init__.py
@@ -1,2 +1,2 @@
-version_info = (1, 6, 1)
+version_info = (1, 7, 0)
 version = '.'.join(str(x) for x in version_info)